Initial check in to master
This is the first check in.
This is the version of Jars that is used in the demo om my linked in profile.

There are still things outstanding like documentation and a few bugs, but this is enough to get it running.

This DOES NOT include a serviceStack or DevExpress Licence as the licences paid for by myself.